[Refactor] #39 - 구조체로 변경, 주석 추가
